biocViews Counts Over Time

biocViews Counts Over Time

Usage

get_view_counts_over_time(view, pkg_list = NULL, pkg_years = NULL)

get_views_counts_over_time(views, pkg_list = NULL, pkg_years = NULL)

Arguments

view

A single biocView term.

pkg_list

Value of a call to biocPkgList(). If NULL (default), will call biocPkgList() internally. See Details.

pkg_years

Value of a call to getPkgYearsInBioc(). If NULL (default), will call getPkgYearsInBioc() internally.

views

A character vector of biocView terms.

Value

For get_view_counts_over_time(), a tibble of two columns: date and count. date is a sequence of dates spaced by six months from 2006 to the current year. count is the number of packages associated with view (or one of its child terms).

For get_views_counts_over_time(), a tibble of $N+1$ columns: one column date and one column for each view in views.

Details

Calling BiocPkgTools::biocPkgList() and passing the result to get_packages_by_view() or get_packages_by_views() is more efficient if you are making multiple calls. See vignette 'Optimisations' for a more comprehensive discussion and demonstration.

Examples

get_view_counts_over_time("SingleCell")
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://p3m.dev/cran/__linux__/noble/latest
#> Checking for Bioc Release Update
#> # A tibble: 40 × 2
#>    date       count
#>    <date>     <int>
#>  1 2006-01-01    13
#>  2 2006-06-01    13
#>  3 2007-01-01    13
#>  4 2007-06-01    13
#>  5 2008-01-01    13
#>  6 2008-06-01    13
#>  7 2009-01-01    14
#>  8 2009-06-01    14
#>  9 2010-01-01    14
#> 10 2010-06-01    14
#> # ℹ 30 more rows
get_views_counts_over_time(c("Spatial", "SingleCell"))
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://p3m.dev/cran/__linux__/noble/latest
#> Checking for Bioc Release Update
#> # A tibble: 40 × 3
#>    date       Spatial SingleCell
#>    <date>       <int>      <int>
#>  1 2006-01-01       4         13
#>  2 2006-06-01       4         13
#>  3 2007-01-01       4         13
#>  4 2007-06-01       4         13
#>  5 2008-01-01       4         13
#>  6 2008-06-01       4         13
#>  7 2009-01-01       4         14
#>  8 2009-06-01       4         14
#>  9 2010-01-01       4         14
#> 10 2010-06-01       4         14
#> # ℹ 30 more rows